Refactoring, updated unit tests to pass with new ai supportability metrics
