+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/Bedrock/Payloads/PayloadInterfaces.cs
@@ -1,9 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-using System.Collections.Generic;
-
-namespace NewRelic.Providers.Wrapper.Bedrock
+namespace NewRelic.Providers.Wrapper.Bedrock.Payloads
 {
     public interface IRequestPayload
     {
